+++ src/edwh_files_plugin/files_plugin.py
@@ -1,86 +1,193 @@
 import json
-import shutil
 import sys
 import tempfile
+import typing
 from pathlib import Path
 from typing import Optional, TypedDict
 
 import requests
-from invoke import Context, task
+from edwh import improved_task as task
+from invoke import Context
 
 # rich.progress is fancier but much slower (100ms import)
 # so use simpler progress library (also used by pip, before rich):
 from progress.bar import ChargingBar
 from requests_toolbelt.multipart.encoder import MultipartEncoder, MultipartEncoderMonitor
-from rich import print
+from rich import print  # noqa: A004
 from termcolor import cprint
 from threadful import thread
 from threadful.bonus import animate
 
+from edwh_files_plugin.compression import DEFAULT_COMPRESSION_LEVEL, Compression
+
 DEFAULT_TRANSFERSH_SERVER = "https://files.edwh.nl"
 
 
-def require_protocol(url: str):
+def require_protocol(url: str) -> str:
     """
     Make sure 'url' has an HTTP or HTTPS schema.
     """
     return url if url.startswith(("http://", "https://")) else f"https://{url}"
 
 
-def create_callback(encoder: MultipartEncoder):
+def create_callback(encoder: MultipartEncoder) -> typing.Callable[[MultipartEncoderMonitor], None]:
+    """
+    Creates a callback function for monitoring the progress of an upload.
+
+    Args:
+        encoder (MultipartEncoder): The multipart encoder that is uploading the data.
+
+    Returns:
+        A callback function that updates a progress bar based on the amount of data that has been read by the encoder.
+
+    Example:
+
+        import requests
+        from requests_toolbelt.multipart.encoder import MultipartEncoder, MultipartEncoderMonitor
+
+        def my_callback(monitor: MultipartEncoder):
+            print('bytes sent: {0}'.format(monitor.bytes_read))
+
+        filename = 'my_file.txt'
+
+        with open(filename, 'rb') as f:
+            encoder = MultipartEncoder(
+                fields={
+                    'file': (filename, f, "text/plain"),
+                }
+            )
+
+        monitor = MultipartEncoderMonitor(encoder, my_callback)
+        result = requests.post('http://some-url.com/upload',
+                               data=monitor,
+                               headers={'Content-Type': monitor.content_type})
+
+    """
     bar = ChargingBar("Uploading", max=encoder.len)
 
-    def callback(monitor: MultipartEncoderMonitor):
+    def callback(monitor: MultipartEncoderMonitor) -> None:
         # goto instead of next because chunk size is unknown
         bar.goto(monitor.bytes_read)
 
     return callback
 
 
-def upload_file(url: str, filename: str, filepath: Path, headers: Optional[dict] = None) -> requests.Response:
+# auto = best for directory; none for files
+# gzip: .tgz for directory; .gz for files. Pigz or Gzip based on availability
+FullCompressionTypes: typing.TypeAlias = typing.Literal["auto", "gzip", "zip", "tgz", "gz", "none"]
+CliCompressionTypes: typing.TypeAlias = typing.Literal["auto", "gzip", "zip", "none"]
+
+
+def upload_file(
+    url: str,
+    filename: str,
+    filepath: Path,
+    headers: Optional[dict[str, typing.Any]] = None,
+    compression: FullCompressionTypes = "auto",
+    compression_level: int = DEFAULT_COMPRESSION_LEVEL,
+) -> requests.Response:
     """
     Upload a file to an url.
     """
     if headers is None:
         headers = {}
 
-    with filepath.open("rb") as f:
-        encoder = MultipartEncoder(
-            fields={
-                filename: (filename, f, "text/plain"),
-            }
-        )
+    with tempfile.TemporaryDirectory() as tmpdir:
+        if compression != "auto":
+            new_filepath = Path(tmpdir) / filepath.name
+            filename = compress_directory(
+                filepath,
+                new_filepath,
+                extension="gz" if compression == "gzip" else compression,
+                compression_level=compression_level,
+            )
+            filepath = new_filepath
 
-        monitor = MultipartEncoderMonitor(encoder, create_callback(encoder))
+        with filepath.open("rb") as f:
+            encoder = MultipartEncoder(
+                fields={
+                    filename: (filename, f, "text/plain"),
+                }
+            )
 
-        return requests.post(url, data=monitor, headers=headers | {"Content-Type": monitor.content_type})  # noqa
+            monitor = MultipartEncoderMonitor(encoder, create_callback(encoder))
+            return requests.post(url, data=monitor, headers=headers | {"Content-Type": monitor.content_type})
 
 
 @thread()
-def _zip_directory(dir_path: str | Path, file_path: str | Path):
+def _compress_directory(
+    dir_path: str | Path,
+    file_path: str | Path,
+    extension: FullCompressionTypes = "auto",
+    compression_level: int = DEFAULT_COMPRESSION_LEVEL,
+) -> str:
     """
-    Compress a directory into a .zip file.
+    Compress a directory into a compressed (zip, gz) file.
     """
-    return shutil.make_archive(str(file_path), "zip", str(dir_path))
+    compressor = Compression.best() if extension == "auto" else Compression.for_extension(extension)
+
+    if not compressor:
+        print(f"[red] No compression available for {extension} [/red]")
+        print(f"[blue] Please choose one of : {Compression.available()}[/blue]")
+        raise RuntimeError("Something went wrong during compression!")
+
+    if compressor.compress(dir_path, file_path, level=compression_level):
+        return compressor.filename(dir_path)
+    else:
+        raise RuntimeError("Something went wrong during compression!")
 
 
-def zip_directory(dir_path: str | Path, file_path: str | Path):
+def compress_directory(
+    dir_path: str | Path,
+    file_path: str | Path,
+    extension: FullCompressionTypes = "auto",
+    compression_level: int = DEFAULT_COMPRESSION_LEVEL,
+) -> str:
     """
-    Compress a directory into a .zip file and show a spinning animation.
+    Compress a directory into a compressed file (zip, gz) and show a spinning animation.
     """
-    return animate(_zip_directory(dir_path, file_path), text=f"Zipping directory {dir_path}")
+    return animate(
+        _compress_directory(dir_path, file_path, extension, compression_level=compression_level),
+        text=f"Compressing directory {dir_path}",
+    )
 
 
-def upload_directory(url: str, filepath: Path, headers: Optional[dict] = None):
+def upload_directory(
+    url: str,
+    filepath: Path,
+    headers: Optional[dict[str, typing.Any]] = None,
+    upload_filename: Optional[str] = None,
+    compression: FullCompressionTypes = "auto",
+    compression_level: int = DEFAULT_COMPRESSION_LEVEL,
+) -> requests.Response:
     """
     Zip a directory and upload it to an url.
+
+    Args:
+        url: which transfer.sh server to use
+        filepath: which directory to upload
+        headers: upload options
+        upload_filename: by default, the directory name with compression extension (e.g. .gz, .zip) will be used
+        compression: which method for compression to use (or best available by default)
+        compression_level: The compression level is a measure of the compression quality (file size; int 0 - 9).
     """
+    filepath = filepath.expanduser().absolute()
     filename = filepath.resolve().name
 
     with tempfile.TemporaryDirectory() as tmpdir:
-        archive_path = zip_directory(filepath, Path(tmpdir) / filename)
+        archive_path = Path(tmpdir) / filename
+        compressed_filename = compress_directory(
+            filepath,
+            archive_path,
+            extension="tgz" if compression == "gzip" else compression,
+            compression_level=compression_level,
+        )  # -> filename.zip e.g.
 
-        return upload_file(url, f"{filename}.zip", Path(archive_path), headers=headers)
+        upload_filename = upload_filename or compressed_filename
+
+        return upload_file(
+            url, upload_filename, Path(archive_path), headers=headers, compression_level=compression_level
+        )
 
 
 class UploadResult(TypedDict):
@@ -97,6 +204,9 @@ def transfer_upload(
     max_downloads: Optional[int] = None,
     max_days: Optional[int] = None,
     encrypt: Optional[str] = None,
+    rename: Optional[str] = None,
+    compression: CliCompressionTypes = "auto",  # auto | pigz | gzip | zip
+    compression_level: int = DEFAULT_COMPRESSION_LEVEL,
 ) -> UploadResult:
     headers: dict[str, str | int] = {}
 
@@ -111,10 +221,28 @@ def transfer_upload(
 
     filepath = Path(filename)
 
-    if filepath.is_dir():
-        response = upload_directory(url, filepath, headers)
-    else:
-        response = upload_file(url, str(filename), filepath, headers)
+    try:
+        if filepath.is_dir():
+            response = upload_directory(
+                url,
+                filepath,
+                headers,
+                upload_filename=rename,
+                compression=compression,
+                compression_level=compression_level,
+            )
+        else:
+            response = upload_file(
+                url,
+                rename or str(filename),
+                filepath,
+                headers,
+                compression=compression,
+                compression_level=compression_level,
+            )
+    except RuntimeError as e:
+        print(f"[red] {e} [/red]")
+        exit(1)
 
     download_url = response.text.strip()
     delete_url = response.headers.get("x-url-delete", "")
@@ -136,7 +264,10 @@ def upload(
     max_downloads: Optional[int] = None,
     max_days: Optional[int] = None,
     encrypt: Optional[str] = None,
-):
+    rename: Optional[str] = None,
+    compression: CliCompressionTypes = "auto",  # auto | pigz | gzip | zip
+    compression_level: int = DEFAULT_COMPRESSION_LEVEL,
+) -> None:
     """
     Upload a file.
 
@@ -147,13 +278,27 @@ def upload(
         max_downloads (int): how often can the file be downloaded?
         max_days (int): how many days can the file be downloaded?
         encrypt (str): encryption password
+        rename (str): upload the file/folder with a different name than it currently has
+        compression (str): by default files are not compressed.
+                           For folders it will try pigz (.tgz), gzip (.tgz) then .zip.
+                           You can also explicitly specify a compression method for files and directory,
+                           and nothing else will be tried.
+        compression_level (int): The compression level is a measure of the compression quality (file size).
+                                 It is expressed as an integer in the range 1 - 9.
+                                 Compression quality and performance are conflicting goals.
+                                 Compression level 1 provides best performance at the expense of quality.
+                                 Compression level 9 provides the smallest file size.
     """
+
     result = transfer_upload(
         filename,
         server=server,
         max_downloads=max_downloads,
         max_days=max_days,
         encrypt=encrypt,
+        rename=rename,
+        compression=compression,
+        compression_level=compression_level,
     )
 
     print(
@@ -165,7 +310,13 @@ def upload(
 
 
 @task(aliases=("get", "receive"))
-def download(_: Context, download_url: str, output_file: Optional[str | Path] = None, decrypt: Optional[str] = None):
+def download(
+    _: Context,
+    download_url: str,
+    output_file: Optional[str | Path] = None,
+    decrypt: Optional[str] = None,
+    unpack: bool = False,
+) -> None:
     """
     Download a file.
 
@@ -174,9 +325,11 @@ def download(_: Context, download_url: str, output_file: Optional[str | Path] = 
         download_url (str): file to download
         output_file (str): path to store the file in
         decrypt (str): decryption token
+        unpack (bool): unpack archive to file(s), removing the archive afterwards
     """
     if output_file is None:
         output_file = download_url.split("/")[-1]
+    output_path = Path(output_file)
 
     download_url = require_protocol(download_url)
 
@@ -184,7 +337,7 @@ def download(_: Context, download_url: str, output_file: Optional[str | Path] = 
     if decrypt:
         headers["X-Decrypt-Password"] = decrypt
 
-    response = requests.get(download_url, headers=headers, stream=True)  # noqa
+    response = requests.get(download_url, headers=headers, stream=True)
 
     if response.status_code >= 400:
         print("[red] Something went wrong: [/red]", response.status_code, response.content.decode(), file=sys.stderr)
@@ -192,20 +345,24 @@ def download(_: Context, download_url: str, output_file: Optional[str | Path] = 
 
     total = int(response.headers["Content-Length"]) // 1024
     with (
-        open(output_file, "wb") as f,
+        output_path.open("wb") as f,
     ):  # <- open file when we're sure the status code is successful!
         for chunk in ChargingBar("Downloading", max=total).iter(response.iter_content(chunk_size=1024)):
             f.write(chunk)
 
+    if unpack:
+        do_unpack(_, str(output_path), remove=True)
+
 
 @task(aliases=("remove",))
-def delete(_: Context, deletion_url: str, quiet: bool = False):
+def delete(_: Context, deletion_url: str, quiet: bool = False) -> bool:
     """
     Delete an uploaded file.
 
     Args:
         _ (Context)
         deletion_url (str): File url + deletion token (from `x-url-delete`, shown in file.upload output)
+        quiet: don't print out server response
     """
     deletion_url = require_protocol(deletion_url)
 
@@ -221,12 +378,40 @@ def delete(_: Context, deletion_url: str, quiet: bool = False):
     return response.ok
 
 
+@task(name="unpack")
+def do_unpack(_: Context, filename: str, remove: bool = False) -> None:
+    """
+    Decompress a given file.
+
+    Args:
+        _: invoke Context
+        filename (str): Path of the file to be decompressed
+        remove (bool, optional): If True, original compressed file will be deleted after decompression.
+                                 Defaults to False.
+
+    Returns:
+        None
+    """
+    filepath = Path(filename)
+    ext = filepath.suffix
+
+    compressor = Compression.for_extension(ext)
+
+    if compressor and compressor.decompress(filepath, filepath.with_suffix("")):
+        if remove:
+            filepath.unlink()
+    else:
+        print("[red] Something went wrong unpacking! [/red]")
+
+
 @task(aliases=("evanescent", "tmp"))
 def evanescence(
     c: Context,
     filename: str | Path,
     server: str = DEFAULT_TRANSFERSH_SERVER,
     encrypt: Optional[str] = None,
+    compression: CliCompressionTypes = "auto",  # auto | pigz | gzip | zip
+    compression_level: int = DEFAULT_COMPRESSION_LEVEL,
 ):
     """
     Temporarily upload a file.
@@ -238,6 +423,8 @@ def evanescence(
         max_days=1,
         max_downloads=2,
         encrypt=encrypt,
+        compression=compression,
+        compression_level=compression_level,
     )
 
     cprint("\nAnd there's just too much that time cannot erase... unlike this file.", on_color="on_cyan")