+++ CHANGELOG.md
@@ -2,6 +2,38 @@
 
 <!--next-version-placeholder-->
 
+## v1.1.1 (2025-03-07)
+
+### Fix
+
+* Show message if `pigz` is not available, warning the user about missing out on performance (once) ([`976b778`](https://github.com/educationwarehouse/edwh-files-plugin/commit/976b778abfeb79a13b7e9a0b6886b1c1a6dba4ce))
+
+## v1.1.0 (2025-02-28)
+
+### Feature
+
+* Allow specifying compression level on `file.upload` ([`511d1b1`](https://github.com/educationwarehouse/edwh-files-plugin/commit/511d1b19a3a5bcdbeececd3d97751c1376ec0bd0))
+
+## v1.0.0 (2025-02-28)
+
+### Feature
+
+* Allow `--compression none` to explicitly send raw or simple tar files ([`46a23ab`](https://github.com/educationwarehouse/edwh-files-plugin/commit/46a23abe703b715ff80684daa58e984c443267f2))
+* Integrate custom (de)compress options with `edwh file` commands ([`8f173fe`](https://github.com/educationwarehouse/edwh-files-plugin/commit/8f173fec844401cc7b8b5bb8daea0f62034dbe1e))
+* Add and test different compression options, todo: implement in actual file upload/download ([`739f05c`](https://github.com/educationwarehouse/edwh-files-plugin/commit/739f05c80ac98631297e39f2b5afc9c3829907f0))
+* WIP on .zip and pigz (.gz) ([`7d25600`](https://github.com/educationwarehouse/edwh-files-plugin/commit/7d25600d974b1f44451992795727458e71a3d2ac))
+
+### Fix
+
+* Better error handling with missing compression ([`6e2dcba`](https://github.com/educationwarehouse/edwh-files-plugin/commit/6e2dcba34b8e9d204be9d8b80581c52d5b486b32))
+* Allow `edwh file.send .` ([`4542a7a`](https://github.com/educationwarehouse/edwh-files-plugin/commit/4542a7af04c06dca560cf8972616e74dde77aa8d))
+* Improved error handling on compress/decompress ([`e02e5af`](https://github.com/educationwarehouse/edwh-files-plugin/commit/e02e5af58a5d1ae6c2c5a13eee555d1d8de44d20))
+
+### Documentation
+
+* Added docstring to compression-related functions ([`7d40d1e`](https://github.com/educationwarehouse/edwh-files-plugin/commit/7d40d1e991d38dbeba0ed78115c1701efbb7659f))
+* Added docstring to compression-related functions ([`5e7241f`](https://github.com/educationwarehouse/edwh-files-plugin/commit/5e7241faa86d0405025900f72ac1e470f63639c3))
+
 ## v0.3.2 (2024-11-21)
 
 ### Fix